+++ project/core/models/day.py
@@ -1,11 +1,11 @@
-from sqlalchemy import BIGINT, Column, DATE
+from sqlalchemy import BigInteger, Column, DATE
 from project.core.models import Base
 
 
 class Day(Base):
     __tablename__ = 'tbl_day'
 
-    id = Column(BIGINT, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     start_at = Column(DATE, nullable=False)
     end_at = Column(DATE, nullable=False)
     eval = Column(BigInteger, nullable=False)
